refactor(solve): extract shared constraint solver (ADR-0011)
Generalize the sketch solver's numeric core into a new top-level solve
package so the 2D sketch engine and the upcoming 3D assembly positioner
(M12-F01) share one solver — the ADR-0011 decision. The damped
Gauss–Newton loop, DOF/redundancy analysis, and dense linear algebra
move verbatim into solve, parameterized over a minimal Residual
interface and []*math.Scalar variables (variable-kind agnostic).

model/sketch/solver.go becomes a thin adapter that keeps the
sketch-facing names (Solve over []Constraint, SolveStatus/SolveResult/
DOFAnalysis as aliases of the solve types), so every sketch call site is
unchanged. The full sketch solver test suite is the regression gate and
stays green.
